Remove splitting executed commands
